requirements.txt
flask
rembg
opencv-python
pillow
numpy
werkzeug